Add Google Nest Device Access component
This is an initial skeleton implementation of the Google Nest
Smart Device Management API.  This currently just gets the authentication
working, without any actual functionality.

This integration uses oauth with a long lived refresh token.

The README includes an overview of how to use the integration, with some
pending work items to complete.